Removed MainPlayer global var. Player instance moved to level class.

+++ Game/src/Weapon.cpp
@@ -105,7 +105,7 @@ bool CWeapon::ModAmmo( int ammo )
 		return false;
 
 	Ammo += ammo;
-	if( Ammo > MaxAmmo )
+	if( Ammo > (int)MaxAmmo )
 		Ammo = MaxAmmo;
 
 	return true;
